requirement.txt
# Existing dependencies
python-dotenv
pvporcupine
pyaudio
livekit
livekit-agents
livekit-plugins-noise-cancellation
livekit-plugins-google
requests
langchain-community
duckduckgo-search

# New dependencies for STT and GPT parsing
SpeechRecognition
openai
pyaudio